feat(website): show resistance mutation annotations on COVID and RSV dashboards
Extract resistance mutation collection configs and helpers from the wasap
subtree into a shared util file (resistanceMutations.ts). Add `byEnv` as
a pure selector to dbIdSpace.ts. Wire COVID, RSV-A, and RSV-B constants
to factory functions that build collection configs; have the generic Astro
page components fetch annotations server-side and pass them down through
the React layout tree.
