+++ website/src/components/views/analyzeSingleVariant/GenericAnalyzeSingleVariantPage.astro
@@ -1,8 +1,16 @@
 ---
+import type { MutationAnnotations } from '@genspectrum/dashboard-components/util';
+
 import { GenericAnalyseSingleVariantReactPage } from './GenericAnalyseSingleVariantReactPage';
-import { isStaging, getDashboardsConfig } from '../../../config';
+import { BackendService } from '../../../backendApi/backendService.ts';
+import { isStaging, getDashboardsConfig, getBackendHost } from '../../../config';
 import BaseLayout from '../../../layouts/base/BaseLayout.astro';
+import { getInstanceLogger } from '../../../logger.ts';
 import { Organisms } from '../../../types/Organism';
+import { getDbIdSpace } from '../../../types/dbIdSpace';
+import { getErrorLogMessage } from '../../../util/getErrorLogMessage';
+import { fetchMutationAnnotationsFromCollections } from '../../../util/resistanceMutations';
+import type { OrganismConstants } from '../../../views/OrganismConstants';
 import { type OrganismWithViewKey } from '../../../views/routing';
 import { ServerSide } from '../../../views/serverSideRouting';
 import { singleVariantViewKey } from '../../../views/viewKeys';
@@ -14,6 +22,23 @@ interface Props {
 const { organism } = Astro.props;
 
 const view = ServerSide.routing.getOrganismView(`${organism}.${singleVariantViewKey}`);
+
+let mutationAnnotations: MutationAnnotations | undefined;
+const constants: OrganismConstants = view.organismConstants;
+const buildCollections = constants.buildResistanceMutationCollections;
+if (buildCollections) {
+    try {
+        const backendService = new BackendService(getBackendHost());
+        mutationAnnotations = await fetchMutationAnnotationsFromCollections(
+            buildCollections(getDbIdSpace()),
+            backendService,
+        );
+    } catch (error) {
+        getInstanceLogger('GenericAnalyzeSingleVariantPage').error(
+            `Failed to fetch resistance mutation annotations for ${organism}: ${getErrorLogMessage(error)}`,
+        );
+    }
+}
 ---
 
 <BaseLayout title={view.viewTitle}>
@@ -21,6 +46,7 @@ const view = ServerSide.routing.getOrganismView(`${organism}.${singleVariantView
         organism={organism}
         organismsConfig={getDashboardsConfig().dashboards.organisms}
         isStaging={isStaging()}
+        mutationAnnotations={mutationAnnotations}
         client:only='react'
     />
 </BaseLayout>
