+++ website/src/components/views/compareSideBySide/GenericCompareSideBySidePage.astro
@@ -1,7 +1,15 @@
 ---
+import type { MutationAnnotations } from '@genspectrum/dashboard-components/util';
+
 import { GenericCompareSideBySideReactPage } from './GenericCompareSideBySideReactPage';
-import { isStaging, getDashboardsConfig } from '../../../config';
+import { BackendService } from '../../../backendApi/backendService.ts';
+import { isStaging, getDashboardsConfig, getBackendHost } from '../../../config';
 import BaseLayout from '../../../layouts/base/BaseLayout.astro';
+import { getInstanceLogger } from '../../../logger.ts';
+import { getDbIdSpace } from '../../../types/dbIdSpace';
+import { getErrorLogMessage } from '../../../util/getErrorLogMessage';
+import { fetchMutationAnnotationsFromCollections } from '../../../util/resistanceMutations';
+import type { OrganismConstants } from '../../../views/OrganismConstants';
 import { type OrganismWithViewKey } from '../../../views/routing';
 import { ServerSide } from '../../../views/serverSideRouting';
 import { compareSideBySideViewKey } from '../../../views/viewKeys';
@@ -15,6 +23,23 @@ const { organism, hideMutationComponents } = Astro.props;
 
 const view = ServerSide.routing.getOrganismView(`${organism}.${compareSideBySideViewKey}`);
 const organismsConfig = getDashboardsConfig().dashboards.organisms;
+
+let mutationAnnotations: MutationAnnotations | undefined;
+const constants: OrganismConstants = view.organismConstants;
+const buildCollections = constants.buildResistanceMutationCollections;
+if (buildCollections) {
+    try {
+        const backendService = new BackendService(getBackendHost());
+        mutationAnnotations = await fetchMutationAnnotationsFromCollections(
+            buildCollections(getDbIdSpace()),
+            backendService,
+        );
+    } catch (error) {
+        getInstanceLogger('GenericCompareSideBySidePage').error(
+            `Failed to fetch resistance mutation annotations for ${organism}: ${getErrorLogMessage(error)}`,
+        );
+    }
+}
 ---
 
 <BaseLayout title={view.viewTitle}>
@@ -23,6 +48,7 @@ const organismsConfig = getDashboardsConfig().dashboards.organisms;
         hideMutationComponents={hideMutationComponents}
         organismsConfig={organismsConfig}
         isStaging={isStaging()}
+        mutationAnnotations={mutationAnnotations}
         client:only='react'
     />
 </BaseLayout>
