+++ website/src/components/views/wasap/wasapPageConfig.ts
@@ -1,6 +1,7 @@
 import type { DateRangeOption, SequenceType, TemporalGranularity } from '@genspectrum/dashboard-components/util';
 
 import type { Organism } from '../../../types/Organism.ts';
+import type { ResistanceMutationCollectionConfig } from '../../../util/resistanceMutations';
 
 export const SEQUENCE_TYPE = {
     nucleotide: 'nucleotide',
@@ -306,12 +307,4 @@ export type WasapFilter = {
     analysis: WasapAnalysisFilter;
 };
 
-/**
- * Resistance mutations defined in a collection, which is specified via the collection ID.
- */
-export type ResistanceMutationCollectionConfig = {
-    collectionId: number;
-    name: string;
-    description: string;
-    annotationSymbol: string;
-};
+export type { ResistanceMutationCollectionConfig } from '../../../util/resistanceMutations';
