+++ website/src/views/covid.ts
@@ -1,4 +1,4 @@
-import { dateRangeOptionPresets, type MutationAnnotation, views } from '@genspectrum/dashboard-components/util';
+import { dateRangeOptionPresets, views } from '@genspectrum/dashboard-components/util';
 
 import {
     getIntegerFromSearch,
@@ -29,19 +29,20 @@ import {
     makeDatasetAndVariantData,
 } from './View.ts';
 import { compareSideBySideViewConstants, singleVariantViewConstants } from './ViewConstants.ts';
+import { buildCovidResistanceMutationCollections } from '../util/resistanceMutations';
 import { CompareSideBySideStateHandler } from './pageStateHandlers/CompareSideBySidePageStateHandler.ts';
 import { type PageStateHandler } from './pageStateHandlers/PageStateHandler.ts';
+import { SingleVariantPageStateHandler } from './pageStateHandlers/SingleVariantPageStateHandler.ts';
 import { setSearchFromDateFilters } from './pageStateHandlers/dateFilterFromToUrl.ts';
+import { setSearchFromLocationFilters } from './pageStateHandlers/locationFilterFromToUrl.ts';
+import { setSearchFromTextFilters } from './pageStateHandlers/textFilterFromToUrl.ts';
+import { advancedQueryUrlParam } from '../components/genspectrum/advancedQueryUrlParamConstants.ts';
+import type { BaselineFilterConfig } from '../components/pageStateSelectors/BaselineSelector.tsx';
 import type { LineageFilterConfig } from '../components/pageStateSelectors/LineageFilterInput.tsx';
 import { Organisms } from '../types/Organism.ts';
 import { type DataOrigin, dataOrigins } from '../types/dataOrigins.ts';
-import { SingleVariantPageStateHandler } from './pageStateHandlers/SingleVariantPageStateHandler.ts';
-import { setSearchFromLocationFilters } from './pageStateHandlers/locationFilterFromToUrl.ts';
-import type { BaselineFilterConfig } from '../components/pageStateSelectors/BaselineSelector.tsx';
 import { ALL_TIMES_LABEL, defaultDateRangeOption } from '../util/defaultDateRangeOption.ts';
 import { formatUrl } from '../util/formatUrl.ts';
-import { setSearchFromTextFilters } from './pageStateHandlers/textFilterFromToUrl.ts';
-import { advancedQueryUrlParam } from '../components/genspectrum/advancedQueryUrlParamConstants.ts';
 
 const earliestDate = '2020-01-06';
 const hostField = 'host';
@@ -127,7 +128,7 @@ class CovidConstants implements OrganismConstants {
     public readonly additionalFilters: Record<string, string> | undefined;
     public readonly dataOrigins: DataOrigin[] = [dataOrigins.nextstrain];
     public readonly accessionDownloadFields = ['strain'];
-    public readonly mutationAnnotations: MutationAnnotation[] = [];
+    public readonly buildResistanceMutationCollections = buildCovidResistanceMutationCollections;
 
     public get aggregatedVisualizations() {
         const hosts = getHostsAggregatedVisualization(this);
